+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Fixed custom analyze script list
+
 ## v0.3.3 (2022-11-30)
 * Switched to Systemic v0.11
 * Updated Dictum dependency